bfm: Move web exporting from sim
