+++ examples/deformation.py
@@ -122,9 +122,9 @@ sim.add_instance(instance)
 sim.add_force(gravity)
 
 sim.run()
-sim.export()
 
 # display results
 # resulting effects from the simulation will automatically be applied to the instance we added to our scene previously
 
+bfm.export(sim)
 bfm.show(sim)
